+++ DX_Training/PoissonDiskDistribution.cpp
@@ -1,4 +1,4 @@
-#include "stdafx.h"
+#include "pch.h"
 
 #include "PoissonDiskDistribution.h"
 
@@ -6,6 +6,8 @@
 #include <cmath>
 
 using namespace DirectX;
+
+// TODO: remove C-style casts in the file
 
 namespace
 {
@@ -46,9 +48,9 @@ void PoissonDiskDistribution::Reset(const DirectX::XMVECTOR& minExtent, const Di
     _gridLengthZ = XMVectorGetZ(_maxExtent) - XMVectorGetZ(_minExtent);
 
     _cellSize = _minSpawnRadius / std::sqrtf(2);
-    _cellsNumX = std::ceil((XMVectorGetX(_maxExtent) - XMVectorGetX(_minExtent)) / _cellSize);
-    _cellsNumY = std::ceil((XMVectorGetY(_maxExtent) - XMVectorGetY(_minExtent)) / _cellSize);
-    _cellsNumZ = std::ceil((XMVectorGetZ(_maxExtent) - XMVectorGetZ(_minExtent)) / _cellSize);
+    _cellsNumX = (size_t)std::ceil((XMVectorGetX(_maxExtent) - XMVectorGetX(_minExtent)) / _cellSize);
+    _cellsNumY = (size_t)std::ceil((XMVectorGetY(_maxExtent) - XMVectorGetY(_minExtent)) / _cellSize);
+    _cellsNumZ = (size_t)std::ceil((XMVectorGetZ(_maxExtent) - XMVectorGetZ(_minExtent)) / _cellSize);
 
     // In the case of a plane, not a volume
     // TODO: bad solution, need something else. "If" for the every coord case... seems to be buggy :P
@@ -117,9 +119,9 @@ DirectX::XMVECTOR PoissonDiskDistribution::GenerateRandomLocationAroundPoint() c
 
 bool PoissonDiskDistribution::CheckCollisions(const DirectX::XMVECTOR& point) const
 {
-    int pointIndexX = ((XMVectorGetX(point) + (_gridLengthX * 0.5f)) / _cellSize);
-    int pointIndexY = ((XMVectorGetY(point) + (_gridLengthY * 0.5f)) / _cellSize);
-    int pointIndexZ = ((XMVectorGetZ(point) + (_gridLengthZ * 0.5f)) / _cellSize);
+    int pointIndexX = (int)((XMVectorGetX(point) + (_gridLengthX * 0.5f)) / _cellSize);
+    int pointIndexY = (int)((XMVectorGetY(point) + (_gridLengthY * 0.5f)) / _cellSize);
+    int pointIndexZ = (int)((XMVectorGetZ(point) + (_gridLengthZ * 0.5f)) / _cellSize);
 
     for (int x = -2; x <= 2; x++)
     {
@@ -154,14 +156,14 @@ bool PoissonDiskDistribution::CheckCollisions(const DirectX::XMVECTOR& point) co
 
 bool PoissonDiskDistribution::AddPointToGrid(const DirectX::XMVECTOR& point, size_t index)
 {
-    int pointIndexX = ((XMVectorGetX(point) + (_gridLengthX * 0.5f)) / _cellSize);
-    int pointIndexY = ((XMVectorGetY(point) + (_gridLengthY * 0.5f)) / _cellSize);
-    int pointIndexZ = ((XMVectorGetZ(point) + (_gridLengthZ * 0.5f)) / _cellSize);
+    int pointIndexX = (int)((XMVectorGetX(point) + (_gridLengthX * 0.5f)) / _cellSize);
+    int pointIndexY = (int)((XMVectorGetY(point) + (_gridLengthY * 0.5f)) / _cellSize);
+    int pointIndexZ = (int)((XMVectorGetZ(point) + (_gridLengthZ * 0.5f)) / _cellSize);
 
     if (_grid[pointIndexX][pointIndexY][pointIndexZ] != -1)
         return false;
 
-    _grid[pointIndexX][pointIndexY][pointIndexZ] = index;
+    _grid[pointIndexX][pointIndexY][pointIndexZ] = (int)index;
 
     return true;
 }
